Move gem name loader to proper namespaced location.
