+++ lib/manageiq/providers/autosde.rb
